Move handle image selected to page
