+++ src/ui/images/pages/mod.rs
@@ -14,4 +14,7 @@ pub use shortcut_images_overview::handle_shortcut_selected;
 
 pub use steam_images_overview::render_page_steam_images_overview;
 pub use select_image_type::render_page_shortcut_select_image_type;
+
 pub use pick_new_image::render_page_pick_image;
+
+pub use pick_new_image::handle_image_selected;
